# src/expenses-api/Api/Resources/Summary/Responses.cs
using System.Collections.Generic;
using System.Runtime.Serialization;
using System.Text.Json.Serialization;

namespace MyAssistant.Apis.Expenses.Api.Resources.Summary
{
    [DataContract]
    public record CurrentSummary
    {
        public string Month { get; set; }
        public string Year { get; set; }
        public decimal TotalAmount { get; set; }
        public Dictionary<string, int> ProgressBar { get; set; } = new Dictionary<string, int>();
    }

    [DataContract]
    public record CompleteSummary
    {
        public string Month { get; set; }
        public string Year { get; set; }
        public decimal Start { get; set; }
        public decimal Saved { get; set; }
        public Dictionary<string, decimal> SpentByCategory { get; set; } = new Dictionary<string, decimal>();
        public Dictionary<string, int> ProgressBar { get; set; } = new Dictionary<string, int>();
    }

    [DataContract]
    public record CategoryBreakdown
    {
        public List<string> Months { get; set; } = new List<string>();
        public List<CategoryRow> Categories { get; set; } = new List<CategoryRow>();
        public decimal TotalAverage { get; set; }
    }

    [DataContract]
    public record CategoryRow
    {
        public string Category { get; set; }
        public List<decimal> MonthlyAmounts { get; set; } = new List<decimal>();
        public decimal Average { get; set; }
    }
}