# src/expenses-api/Api/openapi.yaml
openapi: 3.0.0
info:
  description: ""
  version: "0.3.0"
  title: "Expenses API"
paths:
...
  /api/v1/summary/monthly/{year}/{month}:
    get:
      tags:
        - Summary
      summary: "Get complete summary for a specific month"
      operationId: "GetMonthlySummary"
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
            example: 2025
        - name: month
          in: path
          required: true
          schema:
            type: integer
            example: 7
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteSummary'
...
components:
  schemas:
...
    CompleteSummary:
      type: object
      properties:
        month:
          type: string
          example: July
        year:
          type: string
          example: "2025"
        start:
          type: number
          format: float
          description: Total income (positive amounts)
          example: 2000.00
        saved:
          type: number
          format: float
          description: Net amount (income minus expenses)
          example: 500.00
        spentByCategory:
          type: object
          additionalProperties:
            type: number
            format: float
          description: Category name to absolute amount spent
        progressBar:
          type: object
          additionalProperties:
            type: integer
          description: Category name to percentage of total spending
...
  securitySchemes:
    oAuth:
      type: oauth2
      description: This API uses OAuth 2 with the implicit grant flow.
      flows:
        authorizationCode:
          authorizationUrl: https://api.example.com/oauth2/authorize
          tokenUrl: https://api.example.com/oauth2/token
          scopes:
            read_expenses: read your expenses
            write_expenses: write your expenses